# openapi.yaml
...
openapi: 3.0.3
info:
  title: rucQ API
  version: 0.1.0
  description: rucQ API
servers:
  - url: http://localhost:8080
    description: 開発環境
paths:
...
  /api/camps/{campId}/images:
    get:
      summary: 画像の一覧を取得
      tags:
        - Images
      operationId: getImages
      parameters:
        - $ref: "#/components/parameters/CampId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImageResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/admin/camps/{campId}/images:
    post:
      summary: 画像をアップロード（管理者用）
      tags:
        - Images
      operationId: adminPostImage
      parameters:
        - $ref: "#/components/parameters/CampId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/images/{imageId}:
    get:
      summary: 画像を取得
      tags:
        - Images
      operationId: getImage
      parameters:
        - $ref: "#/components/parameters/ImageId"
      responses:
        "200":
          description: OK
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/admin/images/{imageId}:
    delete:
      summary: 画像を削除（管理者用）
      tags:
        - Images
      operationId: adminDeleteImage
      parameters:
        - $ref: "#/components/parameters/ImageId"
        - $ref: "#/components/parameters/X-Forwarded-User"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    X-Forwarded-User:
      name: X-Forwarded-User
      in: header
      description: ログインしているユーザーのtraQ ID（NeoShowcaseが自動で付与）
      schema:
        type: string
...
    CampId:
      name: campId
      in: path
      description: 合宿ID
      required: true
      schema:
        type: integer
...
    ImageId:
      name: imageId
      in: path
      description: 画像ID
      required: true
      schema:
        type: integer
...
  responses:
...
    NoContent:
      description: No Content
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal server error
  schemas:
...
    ImageResponse:
      type: object
      properties:
        id:
          type: integer
      required:
        - id
...